+++ client/src/features/forecast-load/index.js
@@ -1,2 +1,1 @@
 export { useForecast } from './model/useForecast'
-export { getWeatherIcon } from './model/forecastIcon'
