+++ client/src/features/settlement-search/model/useSettlementSearch.js
@@ -1,4 +1,4 @@
-import { useEffect, useRef, useState } from 'react'
+import { useCallback, useEffect, useRef, useState } from 'react'
 import { searchSettlements } from '@/shared/api'
 
 export function useSettlementSearch() {
@@ -8,7 +8,23 @@ export function useSettlementSearch() {
 	const [error, setError] = useState('')
 	const [isOpen, setIsOpen] = useState(false)
 
+	const containerRef = useRef(null)
 	const skipNextSearchRef = useRef(false)
+	const requestIdRef = useRef(0)
+
+	useEffect(() => {
+		function handleClickOutside(event) {
+			if (!containerRef.current?.contains(event.target)) {
+				setIsOpen(false)
+			}
+		}
+
+		document.addEventListener('mousedown', handleClickOutside)
+
+		return () => {
+			document.removeEventListener('mousedown', handleClickOutside)
+		}
+	}, [])
 
 	useEffect(() => {
 		const trimmedQuery = query.trim()
@@ -28,35 +44,65 @@ export function useSettlementSearch() {
 
 		setIsOpen(true)
 
+		let isCancelled = false
+		const currentRequestId = ++requestIdRef.current
+
 		const timeoutId = setTimeout(async () => {
 			try {
 				setIsLoading(true)
 				setError('')
 
 				const data = await searchSettlements(trimmedQuery)
+
+				if (isCancelled || currentRequestId !== requestIdRef.current) {
+					return
+				}
+
 				setResults(data)
 			} catch {
+				if (isCancelled || currentRequestId !== requestIdRef.current) {
+					return
+				}
+
 				setError('Не удалось выполнить поиск')
 				setResults([])
 			} finally {
+				if (isCancelled || currentRequestId !== requestIdRef.current) {
+					return
+				}
+
 				setIsLoading(false)
 			}
 		}, 300)
 
-		return () => clearTimeout(timeoutId)
+		return () => {
+			isCancelled = true
+			clearTimeout(timeoutId)
+		}
 	}, [query])
 
-	function closeResults() {
-		setIsOpen(false)
+	const clearSearch = useCallback(() => {
+		skipNextSearchRef.current = false
+		setQuery('')
 		setResults([])
 		setError('')
 		setIsLoading(false)
+		setIsOpen(false)
+	}, [])
+
+	function handleFocus() {
+		if (query.trim()) {
+			setIsOpen(true)
+		}
 	}
 
 	function selectSettlement(item) {
 		skipNextSearchRef.current = true
 		setQuery(item.name)
-		closeResults()
+		setIsOpen(false)
+		setResults([])
+		setError('')
+		setIsLoading(false)
 	}
 
 	return {
@@ -66,7 +112,9 @@ export function useSettlementSearch() {
 		isLoading,
 		error,
 		isOpen,
-		closeResults,
+		containerRef,
+		handleFocus,
+		clearSearch,
 		selectSettlement
 	}
 }
