+++ client/src/shared/lib/index.js
@@ -1,3 +1,4 @@
 export { apiClient } from './axios'
 export * from './leaflet/cluster-options'
 export * from './leaflet/marker-icons'
+export * from './forecast/getWeatherIcon'
